Preliminary Database
Created a preliminary database connection class. Currently the database is too simplified to create a Create Script. The conenction is to a MySQL server, allowing the user to enter the appropriate connection to the database if they were to have it. Due to the mysql connector not supporting x86 platforms, x86 support has been removed for the server.