Preliminary Server GUI
This is an overhaul on the server to add a GUI. Currently only conenction to the database and server to start listening for connections are available. All the previous functionality still reamins to be able to lsiten for client connection and broadcast messages back to all the clients. Now there is just a nicer GUI for an administrator to be able to handle the server. More enhancements to come.