Observer Pattern - Observe Server Class
Implemented the observer pattern so classes can observer changes from the server class. This is useful for when messages such as who is connected need to be distributed to multiple windows. Currently only basic implementation has been added to check which host has connected. This will change once the database has been further implemented to log users that are currently conencted, users that disconenct, log messages from users, and more.