Handle Message Type Form Client
Now waits to recieve a first message from the client to determine what kind of message is coming from the client. Currently only the message type is being monitored  to determine that the client has sent a normal message. This is currently logged into the server chat log and eventually will be uploaded to the database when that is completed.