Initial Database And Verifying Login
Create initial database that only includes the current user table. A build script resides in the text file in the parent folder. Additionally, added functionality to receive login credentials from the client and verify they exist in the system. ACK is send back if they are and NACK if they are not.